chore(sdk-core): move api interfaces to sdk-core
Moved api interfaces to sdk-core.

Ticket: BG-48170
