+++ modules/sdk-core/src/index.ts
@@ -1,4 +1,5 @@
 export * from './account-lib';
+export * from './api';
 export * from './bitgo';
 export { BitGoJsError } from './bigojsError';
 import * as common from './common';
